refactor: move puzzle type and generatePuzzle function to puzzle.ts
