+++ web-app/src/components/game/game.tsx
@@ -1,6 +1,7 @@
 import { cn } from "@/lib/utils";
-import { Cell, Puzzle } from "@/shared/common";
+import { Cell } from "@/shared/common";
 import { DIGITS, Digit } from "@/shared/digit";
+import { Puzzle } from "@/shared/puzzle";
 import { useEffect, useReducer, useState } from "react";
 import { cellSiblings, gameReducer, setup } from "./game_reducer";
 import { ToggleableButton } from "./toggleable_button";
