+++ web-app/src/components/home.tsx
@@ -1,8 +1,8 @@
 import { Game } from "./game/game";
-import { generatePuzzle } from "./game/game_reducer";
+import { generatePuzzle } from "@/shared/puzzle";
 
 function Home() {
-  return <Game puzzle={generatePuzzle("Hard")} />;
+  return <Game puzzle={generatePuzzle("Medium", 0)} />;
 }
 
 export { Home };
